backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-WTF==1.2.1
WTForms==3.1.0
Werkzeug==3.0.1

# Database
psycopg[binary]==3.2.1
SQLAlchemy==2.0.23

# File handling and validation
Pillow==10.4.0
pandas==2.1.3
openpyxl==3.1.2
PyPDF2==3.0.1

# Geolocation and mapping
requests==2.31.0
folium==0.15.0

# AI/ML
scikit-learn==1.3.2
numpy==1.25.2
openai==1.3.7

# Email and notifications
Flask-Mail==0.9.1

# PDF generation
reportlab==4.0.7

# Environment variables
python-dotenv==1.0.0

# Authentication and security
Flask-Bcrypt==1.0.1
PyJWT==2.8.0

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
flake8==6.1.0
black==23.11.0
gunicorn==21.2.0